+++ src/Resources/PaymentTypes/InstallmentSecured.php
@@ -20,11 +20,15 @@
  *
  * @package  UnzerSDK\PaymentTypes
  */
+
 namespace UnzerSDK\Resources\PaymentTypes;
 
 use DateTime;
 use UnzerSDK\Resources\InstalmentPlan;
 
+/** @deprecated will be replaced by PaylaterInstallment.
+ * @see PaylaterInstallment
+ */
 class InstallmentSecured extends InstalmentPlan
 {
     /** @var string $iban */
@@ -37,11 +41,11 @@ class InstallmentSecured extends InstalmentPlan
     protected $accountHolder;
 
     /**
-     * @param InstalmentPlan|null  $selectedPlan
-     * @param null|string          $iban
-     * @param null|string          $accountHolder
+     * @param InstalmentPlan|null $selectedPlan
+     * @param null|string $iban
+     * @param null|string $accountHolder
      * @param null|DateTime|string $orderDate
-     * @param null|string          $bic
+     * @param null|string $bic
      * @param null|DateTime|string $invoiceDate
      * @param null|DateTime|string $invoiceDueDate
      */
@@ -49,8 +53,8 @@ class InstallmentSecured extends InstalmentPlan
     {
         parent::__construct();
 
-        $this->iban          = $iban;
-        $this->bic           = $bic;
+        $this->iban = $iban;
+        $this->bic = $bic;
         $this->accountHolder = $accountHolder;
         $this->setOrderDate($orderDate);
         $this->setInvoiceDate($invoiceDate);